Renamed project and packages
Used recommended conventions from the Dropwizard style guide.
