+++ UrlShortener/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>UrlShortener</name>
+	<name>shrtn</name>
 	<comment></comment>
 	<projects>
 	</projects>
